+++ nitpicker/cvs/__init__.py
@@ -1,0 +1,1 @@
+from .cvs_factory import CVSFactory